% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_table.R
\name{author_table}
\alias{author_table}
\title{Get author information in a table}
\usage{
author_table(paper)
}
\arguments{
\item{paper}{a paper object or a list of paper objects}
}
\value{
a data frame of author information
}
\description{
Get author information in a table
}
\examples{
paper <- psychsci[1:2]
author_table(paper)
}
