% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-helpers.R
\name{collapse_section}
\alias{collapse_section}
\title{Make Collapsible Section}
\usage{
collapse_section(
  text,
  title = "Learn More",
  callout = c("tip", "note", "warning", "important", "caution"),
  collapse = TRUE
)
}
\arguments{
\item{text}{The text to put in the collapsible section; vectors will be collapse with line breaks between (e.g., into paragraphs)}

\item{title}{The title of the collapse header}

\item{callout}{the type of quarto callout block}

\item{collapse}{whether to collapse the block at the start}
}
\value{
text
}
\description{
A helper function for making module reports.
}
\examples{
text <- c("Paragraph 1...", "Paragraph 2...")
collapse_section(text) |> cat()
}
