% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{compare_tables}
\alias{compare_tables}
\title{Compare Tables for Validation}
\usage{
compare_tables(
  expected,
  observed,
  match_cols = c("id", "text"),
  comp_cols = NULL
)
}
\arguments{
\item{expected}{the expected table}

\item{observed}{the observed table}

\item{match_cols}{which columns should be used to determine identification}

\item{comp_cols}{which columns should be compared for classification}
}
\value{
a list of comparisons
}
\description{
Compare Tables for Validation
}
\examples{
expected <- data.frame(id = 1:2, text = c("A", "B"), value = c(10, 20))
observed <- data.frame(id = 1:2, text = c("A", "B"), value = c(10, 25))
compare_tables(expected, observed)
}
