% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{crossref_doi}
\alias{crossref_doi}
\title{CrossRef Info from DOI}
\usage{
crossref_doi(
  doi,
  select = c("DOI", "type", "title", "container-title", "volume", "issue", "page", "URL",
    "abstract", "year", "error")
)
}
\arguments{
\item{doi}{the DOI of the paper to get info for}

\item{select}{what fields to select from the crossref API}
}
\value{
data frame with DOIs and info
}
\description{
Valid selects for crossref API are:
}
\details{
abstract, URL, resource, member, posted, score, created, degree, update-policy, short-title, license, ISSN, container-title, issued, update-to, issue, prefix, approved, indexed, article-number, clinical-trial-number, accepted, author, group-title, DOI, is-referenced-by-count, updated-by, event, chair, standards-body, original-title, funder, translator, published, archive, published-print, alternative-id, subject, subtitle, published-online, publisher-location, content-domain, reference, title, link, type, publisher, volume, references-count, ISBN, issn-type, assertion, deposited, page, content-created, short-container-title, relation, editor
}
\examples{
doi <- "10.7717/peerj.4375"
\dontrun{
# cr_info <- crossref_doi(doi)
}
}
