% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{crossref_query}
\alias{crossref_query}
\title{Look up Reference in CrossRef}
\usage{
crossref_query(
  ref,
  min_score = 50,
  rows = 1,
  select = c("DOI", "score", "type", "title", "author", "container-title", "published",
    "volume", "issue", "page", "URL", "abstract")
)
}
\arguments{
\item{ref}{the full text reference of the paper to get info for, see Details}

\item{min_score}{minimal score that is taken to be a reliable match (default 50)}

\item{rows}{the maximum number of rows to return per reference (default 1)}

\item{select}{what fields to select from the crossref API}
}
\value{
doi
}
\description{
Look up Reference in CrossRef
}
\details{
The argument \code{ref} can take many formats.  Crossref queries only look for authors, title, and container-title (e.g., journal or book), but extra information doesn't seem to hurt.
\itemize{
\item be a text reference or fragment
\item a bibentry object (authors, title and container will be extracted)
\item a vector of text or bibentry objects
\item a paper object (the ref column of the bib table will be extracted)
}
}
\examples{
ref <- paste(
  "Lakens, D., Mesquida, C., Rasti, S., & Ditroilo, M. (2024).",
  "The benefits of preregistration and Registered Reports.",
  "Evidence-Based Toxicology, 2(1)."
)
\donttest{
  cr <- crossref_query(ref)
}
}
