% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{On Attach}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{libname}

\item{pkgname}{pkgname}
}
\value{
startup message
}
\description{
On Attach
}
\keyword{internal}
