% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-extractors.R
\name{extract_urls}
\alias{extract_urls}
\title{Extract URLs}
\usage{
extract_urls(paper)
}
\arguments{
\item{paper}{a paper object or paperlist object}
}
\value{
a table
}
\description{
Get a table of URLs from a paper or paperlist. Matches urls that start with http or doi:
}
\examples{
paper <- read(demoxml())
urls <- extract_urls(paper)
}
