% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-helpers.R
\name{format_ref}
\alias{format_ref}
\title{Format Reference}
\usage{
format_ref(bib)
}
\arguments{
\item{bib}{a bibentry object or list of bibentry objects}
}
\value{
formatted text
}
\description{
Format a reference for display in a report.
}
\details{
The argument \code{bib} should be a bibentry object (e.g., like those made by \code{citation()}, but it can also handle a bibtex object or a bibtex formatted character vector. If these do not read in as valid bibtex, the original text of bib will be returned unformatted.
}
\examples{
mc <- citation("metacheck")
format_ref(mc)

# handles bibtext
bib_mc <- utils::toBibtex(mc)
format_ref(bib_mc)

paper <- read(demoxml())
format_ref(paper$bib$ref[1:2])
}
