% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{get_prev_outputs}
\alias{get_prev_outputs}
\title{Get Previous Outputs}
\usage{
get_prev_outputs(module, item)
}
\arguments{
\item{module}{the name of a previously run module}

\item{item}{the name of the list item to extract}
}
\value{
the extracted list item, or NULL if not found
}
\description{
A helper for creating modules. Checks for previous module outputs in a chain and returns the named list item if it exists in any parent environment.
}
\examples{
# .prev_outputs__ is usually created by `module_run()`
.prev_outputs__ <- list(mod_1 = list(a = 1, b = 2))
f <- function(item) { get_prev_outputs("mod_1", item) }
f("a")
f("d")
}
