% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_paper_list}
\alias{is_paper_list}
\title{Detect a list of paper objects}
\usage{
is_paper_list(paper)
}
\arguments{
\item{paper}{the object to test}
}
\value{
logical
}
\description{
Detect a list of paper objects
}
\keyword{internal}
