% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-helpers.R
\name{link}
\alias{link}
\title{Make an html link}
\usage{
link(url, text = url, new_window = TRUE, type = "")
}
\arguments{
\item{url}{the URL to link to}

\item{text}{the text to link}

\item{new_window}{whether to open in a new window}

\item{type}{handle common links, like "doi" ()}
}
\value{
string
}
\description{
Make an html link
}
\examples{
link("https://scienceverse.org")
}
