% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm}
\alias{llm}
\title{Query an LLM}
\usage{
llm(
  text,
  system_prompt,
  text_col = "text",
  model = llm_model(),
  params = list()
)
}
\arguments{
\item{text}{The text to send to the LLM (vector of strings, or data frame with the text in a column)}

\item{system_prompt}{A system prompt to set the behavior of the assistant}

\item{text_col}{The name of the text column if text is a data frame}

\item{model}{the LLM model name (see \code{llm_model_list()}) in the format "provider" or "provider/model"}

\item{params}{a named list to pass to \code{ellmer::params()}}
}
\value{
a list of results
}
\description{
Ask a large language model (LLM) any question you want about a vector of text or the text from a search_text().
}
\details{
You will need to get your own API key from \url{https://console.groq.com/keys}. To avoid having to type it out, add it to the .Renviron file in the following format (you can use \code{usethis::edit_r_environ()} to access the .Renviron file)

GROQ_API_KEY="key_value_asdf"

See \url{https://console.groq.com/docs} for more information
}
\examples{
\dontrun{
  text <- c("hello", "number", "ten", 12)
  system_prompt <- "Is this a number? Answer only 'TRUE' or 'FALSE'"
  is_number <- llm(text, system_prompt)
  is_number
}
}
