% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_model_list}
\alias{llm_model_list}
\title{List Available LLM Models}
\usage{
llm_model_list(API_KEY = Sys.getenv("GROQ_API_KEY"))
}
\arguments{
\item{API_KEY}{groq API key from \url{https://console.groq.com/keys}}
}
\value{
a data frame of models and info
}
\description{
Returns a list of available models in groq, excluding whisper and vision models (for audio and images) and sorting by creation date. See \url{https://console.groq.com/docs/models} for more information.
}
\examples{
\donttest{
  llm_model_list()
}
}
