% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{module_report}
\alias{module_report}
\title{Report from module output}
\usage{
module_report(module_output, header = 3)
}
\arguments{
\item{module_output}{the output of a \code{module_run()}}

\item{header}{header level (default 2)}
}
\value{
text
}
\description{
Report from module output
}
\examples{
filename <- demoxml()
paper <- read(filename)
op <- module_run(paper, "stat_p_exact")
module_report(op) |> cat()
}
