% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{openalex_query}
\alias{openalex_query}
\title{Look up a reference in OpenAlex}
\usage{
openalex_query(title, source = NA, authors = NA, strict = TRUE)
}
\arguments{
\item{title}{The title of the work}

\item{source}{The source (journal or book)}

\item{authors}{The authors}

\item{strict}{Whether to return NULL or the best match if there isn't a single match}
}
\value{
A data frame with citation info
}
\description{
Look up a reference in OpenAlex
}
\examples{
\dontrun{
openalex_query("Sample Size Justification", "Collabra Psychology")
}
}
