% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-helpers.R
\name{plural}
\alias{plural}
\title{Pluralise}
\usage{
plural(n, singular = "", plural = "s")
}
\arguments{
\item{n}{the number}

\item{singular}{the word or ending when n = 1}

\item{plural}{the word or ending n != 1}
}
\value{
a string
}
\description{
Helper function for conditional plurals. For example, if you want to return "1 error" or "2 errors", you can use this in a sprintf().
}
\examples{
n <- 0:3
sprintf("I have \%d friend\%s", n, plural(n))
sprintf("I have \%d \%s", n, plural(n, "octopus", "octopi"))
}
