% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scienceverse.R
\name{print.scivrs_author}
\alias{print.scivrs_author}
\title{Print Author}
\usage{
\method{print}{scivrs_author}(x, ...)
}
\arguments{
\item{x}{The scivrs_author item}

\item{...}{Additional parameters for print}
}
\description{
Print Author
}
\keyword{internal}
