% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubpeer.R
\name{pubpeer_comments}
\alias{pubpeer_comments}
\title{Get Pubpeer Comments}
\usage{
pubpeer_comments(doi)
}
\arguments{
\item{doi}{a vector of paper DOIs}
}
\value{
a dataframe with information from pubpeer
}
\description{
Takes a DOI, and retrieves information from pubpeer related to post-publication peer review comments.
}
\examples{
doi <- c(
  "10.1038/s41598-025-24662-9",
  "10.1177/0146167211398138"
)
pubpeer_comments(doi)
}
