% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2grobid.R
\name{pytacheck}
\alias{pytacheck}
\title{Process a paper using the Pytacheck API}
\usage{
pytacheck(
  file_path,
  api_url = "https://api.metacheck.app/get-paper-metadata/",
  api_key = Sys.getenv("PYTACHECK_API")
)
}
\arguments{
\item{file_path}{Path to the PDF file}

\item{api_url}{Base URL of the API}

\item{api_key}{Key to access pytacheck}
}
\value{
A list of parsed information
}
\description{
Process a paper using the Pytacheck API
}
\keyword{internal}
