% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchbox.R
\name{rbox_file_download}
\alias{rbox_file_download}
\title{Retrieve files from ResearchBox by URL}
\usage{
rbox_file_download(rb_url)
}
\arguments{
\item{rb_url}{a vector of ResearchBox URLs}
}
\value{
a data frame of information
}
\description{
Retrieve files from ResearchBox by URL
}
\keyword{internal}
