% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{crossref_doi}
\alias{crossref_doi}
\title{CrossRef Info from DOI}
\usage{
crossref_doi(doi)
}
\arguments{
\item{doi}{the DOI of the paper to get info for}
}
\value{
crossref data
}
\description{
CrossRef Info from DOI
}
\examples{
doi <- "10.7717/peerj.4375"
\dontrun{
 # cr_info <- crossref_doi(doi)
}
}
