% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{crossref_query}
\alias{crossref_query}
\title{Look up Reference in CrossRef}
\usage{
crossref_query(ref, min_score = 50, rows = 1)
}
\arguments{
\item{ref}{the full text reference of the paper to get info for}

\item{min_score}{minimal score that is taken to be a reliable match (default 50)}

\item{rows}{the maximum number of rows to return}
}
\value{
doi
}
\description{
Look up Reference in CrossRef
}
\examples{
ref <- paste(
  "Lakens, D., Mesquida, C., Rasti, S., & Ditroilo, M. (2024).",
  "The benefits of preregistration and Registered Reports.",
  "Evidence-Based Toxicology, 2(1)."
)
\donttest{
  cr <- crossref_query(ref)
}
}
