% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_model_list}
\alias{llm_model_list}
\title{List LLM Models}
\usage{
llm_model_list(platform = NULL)
}
\arguments{
\item{platform}{The platform. If NULL, checks all platforms for which you have a valid API_KEY.}
}
\value{
a data frame of models and info
}
\description{
List available LLM models for the specified platform.
}
\details{
For platforms other than groq, returns the value from the corresponding ellmer::models_platform function.
}
\examples{
\dontrun{
  llm_model_list()
}
}
