% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2grobid.R
\name{pytacheck}
\alias{pytacheck}
\title{Process a paper using the Pytacheck API}
\usage{
pytacheck(file_path, api_url = "http://api.metacheck.app")
}
\arguments{
\item{file_path}{Path to the PDF file}

\item{api_url}{Base URL of the API}
}
\value{
A Tibble containing the processed data (id, filename, grobid_xml, ocr_markdown, etc.)
}
\description{
Process a paper using the Pytacheck API
}
\keyword{internal}
