% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-helpers.R
\name{scroll_table}
\alias{scroll_table}
\title{Make Scroll Table}
\usage{
scroll_table(
  table,
  colwidths = "auto",
  maxrows = 2,
  escape = FALSE,
  column = "body"
)
}
\arguments{
\item{table}{the data frame to show in a table, or a vector for a list}

\item{colwidths}{set column widths as a vector of px (number > 1) or percent (numbers <= 1)}

\item{maxrows}{if the table has more rows than this, paginate}

\item{escape}{whether or not to escape the DT (necessary if using raw html)}

\item{column}{which quarto column to show tables in}
}
\value{
the markdown R chunk to create this table
}
\description{
A helper function for making module reports.
}
\details{
See \href{https://quarto.org/docs/authoring/article-layout.html}{quarto article layout} for column options. The most common are "body" (centre column), "page" (span all columns"), and "margin" (only in right margin).

To set colwidths, use a numeric or character vector. For a numeric vector, numbers greater than 1 will be interpreted as pixels, less than 1 as percents. Character vectors will be passed as is (e.g., "3em"). If you only want to specify some columns, set the others to NA, like c(200, NA, 200, NA).
}
\examples{
scroll_table(LETTERS)
}
